+++ src/de/var.rs
@@ -38,7 +38,9 @@ where
             DeEvent::Text(t) => EscapedDeserializer::new(Cow::Borrowed(t), decoder, true),
             // Escape sequences does not processed inside CDATA section
             DeEvent::CData(t) => EscapedDeserializer::new(Cow::Borrowed(t), decoder, false),
-            DeEvent::Start(e) => EscapedDeserializer::new(Cow::Borrowed(e.name()), decoder, false),
+            DeEvent::Start(e) => {
+                EscapedDeserializer::new(Cow::Borrowed(e.name().into_inner()), decoder, false)
+            }
             _ => {
                 return Err(DeError::Unsupported(
                     "Invalid event for Enum, expecting `Text` or `Start`",
@@ -65,7 +67,7 @@ where
 
     fn unit_variant(self) -> Result<(), DeError> {
         match self.de.next()? {
-            DeEvent::Start(e) => self.de.read_to_end(e.name()),
+            DeEvent::Start(e) => self.de.read_to_end(e.name().as_ref()),
             DeEvent::Text(_) | DeEvent::CData(_) => Ok(()),
             _ => unreachable!(),
         }